Tests/LibWeb/Text/input/HTML/img-element-fetch-priority.html

<img id="testImg">
<script src="../include.js"></script>
<script>
    test(() => {
        const img = document.getElementById('testImg');

        println(`fetchPriority = ${img.fetchPriority}`);
        println('');

        img.fetchPriority = 'high';

        println(`fetchPriority = ${img.fetchPriority}`);
        println('');

        img.fetchPriority = 'invalid-value';
        println(`fetchPriority = ${img.fetchPriority}`);
    });
</script>
