Tests/LibWeb/Text/expected/HTML/img-element-fetch-priority.txt

   fetchPriority = auto

fetchPriority = high

fetchPriority = auto
